Move MultilineTextBox to widget
